+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 0.4.0
 ### Features
 * Allowed configurable EL & CL client images via the `elImage` and `clImage` keys to participant object
 * Support `trace` loglevel